Working revision
In this revision, flooding of phishing pages works like a charm. However
it's not distributed to a network of applications what makes it easy for
the Phisher to filter the flooded data as it's all generated by a single
IP.

A P2P system, or kind of a botnet, has to be implemented which takes
care of this issue by providing a larger amount of sources from which
the data is flooded into the Phisher's database.